requirements.txt
Django==2.2.12
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==7.1.1
pytz==2019.3
screen==1.0.1
sqlparse==0.3.1
xlwt==1.3.0
